backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
anthropic==0.39.0
httpx==0.27.2
azure-cognitiveservices-speech==1.40.0
edge-tts==7.2.8
